requirements.txt
sqlalchemy~=2.0
psycopg2-binary~=2.9.10
pandas~=2.2
requests~=2.32
beautifulsoup4~=4.12
google-auth~=2.36
google-api-python-client ~=2.152
protobuf~=6.30.2
pytest-cov ~=6.0
dotenv~=0.9.9
